Prototype a callback for writing
I can't seem to find a good way to get the data from curl instead
of printing to STDOUT. This is an attempt to use a buffer in C and
a function that returns what's in that buffer. It makes the userp
parameter in the callback useless but I don't think there is a way
to change arguments in Stanza.

This commit also contains a file for producing ctags.